Moved code to first opencast js module
