+++ js/opencast/rollup.config.js
@@ -1,0 +1,20 @@
+import {terser} from 'rollup-plugin-terser';
+
+export default {
+    external: [
+        'document',
+        'ilias',
+        'jquery',
+    ],
+    input: './src/index.js',
+    output: {
+        file: './dist/index.js',
+        format: 'iife',
+        globals: {
+            document: 'document',
+            ilias: 'il',
+            jquery: '$',
+        }
+    },
+    plugins: [terser()]
+};
