+++ js/opencast/src/index.js
@@ -1,0 +1,7 @@
+import il from 'ilias';
+import $ from 'jquery';
+import PasswordToggle from './Form/PasswordToggle';
+
+il.Opencast = il.Opencast || {};
+il.Opencast.Form = il.Opencast.Form || {};
+il.Opencast.Form.passwordToggle = new PasswordToggle($);
